Added OverlayFS cleanup worker
The OverlayFS now periodically checks all sandboxes for the presence of
work-... and merge-... dirs which are not in use by the current ossh
process. If it finds any it will attempt to unmount the OverlayFS and
remove the directories. This should cleanup any unclosed directories on
startup as well as adding a automatic retry to the FS closing in case
the FS is still flushing or busy for some other reason